Rename api_config.rs to features.rs
