+++ clis/tcli/src/cli.rs
@@ -5,8 +5,8 @@ use parking_lot::Mutex;
 use serde::{Deserialize, Serialize};
 use telio::crypto::{PublicKey, SecretKey};
 use telio::device::{Device, DeviceConfig};
-use telio_model::api_config::Features;
 use telio_model::config::{RelayState, Server};
+use telio_model::features::Features;
 use telio_model::{config::Config as MeshMap, event::Event as DevEvent, mesh::ExitNode};
 use telio_proto::{CodecError, PacketTypeRelayed};
 use telio_wg::AdapterType;
